Split video.py
